+++ modules/libasio/asio_node.h
@@ -97,7 +97,7 @@ namespace module
 				//@e [out] : 错误码
 				virtual void afterPolledSendDataNotification(
 					const uint32_t id = 0, 
-					const uint32_t bytes = 0, 
+					const uint64_t bytes = 0, 
 					const int32_t e = 0) = 0;
 			};//class ASIONode
 		}//namespace asio
